requirements.txt
rake-nltk
pytextrank
appdirs==1.4.4
attrs==19.3.0
black==19.10b0
blis==0.4.1
boto3==1.12.37
botocore==1.15.37
cachetools==4.0.0
catalogue==1.0.0
certifi>=2020.06.20
chardet==3.0.4
click==7.1.1
cymem==2.0.3
dash==1.10.0
dash-bootstrap-components==0.9.2
dash-core-components==1.9.0
dash-cytoscape==0.1.1
dash-html-components==1.0.3
dash-renderer==1.3.0
dash-table==4.6.2
docutils==0.15.2
Flask==1.1.2
Flask-Caching==1.8.0
Flask-Compress==1.4.0
funcy==1.14
future==0.18.2
gensim==3.8.1
google
google-api-core==1.16.0
google-auth==1.13.1
google-cloud-core==1.3.0
google-cloud-storage==1.27.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.14.1
llvmlite==0.31.0
MarkupSafe==1.1.1
more-itertools==8.2.0
murmurhash==1.0.2
numba==0.48.0
numexpr==2.7.1
numpy==1.18.2
packaging==20.3
pandas==1.0.3
pathspec==0.8.0
plac==1.1.3
plotly==4.6.0
pluggy==0.13.1
preshed==3.0.2
protobuf==3.11.3
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyLDAvis==2.1.2
pyparsing==2.4.7
pytest==5.4.1
python-dateutil==2.8.1
pytz==2019.3
regex==2020.5.14
requests==2.23.0
retrying==1.3.3
rsa==4.0
s3transfer==0.3.3
scikit-learn==0.20.1
scipy==1.4.1
six==1.14.0
sklearn==0.0
smart-open==1.10.0
spacy==2.2.4
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz#egg=en_core_web_sm
srsly==1.0.2
tbb==2019.0
thinc==7.4.0
threadpoolctl==2.0.0
toml==0.10.1
tqdm==4.45.0
typed-ast==1.4.1
umap-learn==0.4.0
urllib3==1.25.8
wasabi==0.6.0
wcwidth==0.1.9
Werkzeug==1.0.1
zipp==3.1.0
lxml
wordcloud
